test: move common mocks into namespace
