+++ src/__tests__/__mocks__/fs/index.ts
@@ -1,0 +1,3 @@
+import { fs } from 'memfs';
+
+module.exports = fs;
