+++ src/electron/requests/storage/ElectronFilesController.integration.test.ts
@@ -1,67 +1,12 @@
-import type { app } from 'electron';
-import memfsNameSpace, { vol } from 'memfs';
-import * as electronIPCMock from '@electron/utils/ipc/__tests__/electron';
-import { getResolvedPath, joinPathSegments } from '@utils/fs/paths';
+import { vol } from 'memfs';
 
 import { enableStorage } from './main';
 import { ElectronFilesController, storageApi } from './renderer';
 
-// Mock API to prevent writes to a real FS
-vi.mock('fs', async () => {
-	// importActual so types/impl come from memfs
-	const memfs = await vi.importActual('memfs');
-	// return the promises API so code that does `import { promises as fs } from 'node:fs/promises'`
-	// will receive memfs.fs.promises
-	return memfs.fs as typeof memfsNameSpace;
-});
-vi.mock('fs/promises', async () => {
-	const memfs = await vi.importActual('memfs');
-	return (memfs as typeof memfsNameSpace).fs.promises;
-});
-
-vi.mock('electron', async () => {
-	return {
-		app: {
-			// We run tests against production code
-			isPackaged: true,
-			getPath: ((name) => {
-				switch (name) {
-					case 'userData':
-						return '/home/userData/appDir';
-					default:
-						throw new Error(`Not implemented case for "${name}"`);
-				}
-			}) satisfies (typeof app)['getPath'],
-		},
-		...electronIPCMock,
-	};
-});
-
-// Real package uses a system calls, that is not mocked,
-// so we mock behaviour of this package
-vi.mock('recursive-readdir', async () => {
-	return {
-		default: async (path: string) => {
-			const results = await vol.promises.readdir(path, { recursive: true });
-			return (
-				results
-					.map((foundPath) =>
-						joinPathSegments([
-							getResolvedPath(path, '/'),
-							foundPath as string,
-						]),
-					)
-					// Filter out directories and leave only files
-					.filter((path) => {
-						// Standard `isDirectory` api implemented incorrect,
-						// so we manually implement check for directory
-						const isDirectory = vol.statSync(path).mode === 16895;
-						return !isDirectory;
-					})
-			);
-		},
-	};
-});
+vi.mock('fs', () => vi.importActual('@mocks/fs'));
+vi.mock('fs/promises', () => vi.importActual('@mocks/fs/promises'));
+vi.mock('recursive-readdir', () => vi.importActual('@mocks/recursive-readdir'));
+vi.mock('electron', () => vi.importActual('@mocks/electron'));
 
 const storageHandlerCleanup = enableStorage();
 const filesController = new ElectronFilesController(storageApi, 'profileDir');
